Move tensor evaluation code to test-only
