+++ pytket/tests/zx_tensor/__init__.py
@@ -17,4 +17,13 @@ Python API module for interfacing with tket c++ implemented functions and module
 Exports class ZXDiagram and associated tools.
 """
 
-from pytket._tket.zx import *  # type: ignore
+from .tensor_eval import (
+    tensor_from_quantum_diagram,
+    tensor_from_mixed_diagram,
+    unitary_from_classical_diagram,
+    unitary_from_quantum_diagram,
+    density_matrix_from_cptp_diagram,
+    fix_boundaries_to_binary_states,
+    fix_inputs_to_binary_state,
+    fix_outputs_to_binary_state,
+)
